Move components into /components/external/*
- Do the same for __tests__
- Add "external" to storybook titles
