+++ src/__tests__/external/BreadCrumb.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, mount, render } from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
-import BreadCrumb from '../components/BreadCrumb';
+import BreadCrumb from '../../components/external/BreadCrumb';
 
 // Shallow Rendering
 // https://github.com/airbnb/enzyme/blob/master/docs/api/shallow.md
