+++ src/__tests__/external/NavBar.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, mount, render } from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
-import NavBar from '../components/NavBar';
+import NavBar from '../../components/external/NavBar';
 
 // Shallow Rendering
 // https://github.com/airbnb/enzyme/blob/master/docs/api/shallow.md
