+++ src/__tests__/external/TitleAndDescription.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow, mount, render } from 'enzyme';
 import { expect } from 'chai';
 import sinon from 'sinon';
-import TitleAndDescription from '../components/TitleAndDescription';
+import TitleAndDescription from '../../components/external/TitleAndDescription';
 
 // Shallow Rendering
 // https://github.com/airbnb/enzyme/blob/master/docs/api/shallow.md
